+++ src/components.d.ts
@@ -453,7 +453,16 @@ export namespace Components {
          */
         "qrCodeSvg": string;
     }
+    interface MvxWalletConnectAppGalleryIcon {
+        "class"?: string;
+    }
+    interface MvxWalletConnectAppStoreIcon {
+        "class"?: string;
+    }
     interface MvxWalletConnectDownload {
+        "class"?: string;
+    }
+    interface MvxWalletConnectGooglePlayIcon {
         "class"?: string;
     }
     interface MvxWalletConnectScan {
@@ -1164,11 +1173,29 @@ declare global {
         prototype: HTMLMvxWalletConnectElement;
         new (): HTMLMvxWalletConnectElement;
     };
+    interface HTMLMvxWalletConnectAppGalleryIconElement extends Components.MvxWalletConnectAppGalleryIcon, HTMLStencilElement {
+    }
+    var HTMLMvxWalletConnectAppGalleryIconElement: {
+        prototype: HTMLMvxWalletConnectAppGalleryIconElement;
+        new (): HTMLMvxWalletConnectAppGalleryIconElement;
+    };
+    interface HTMLMvxWalletConnectAppStoreIconElement extends Components.MvxWalletConnectAppStoreIcon, HTMLStencilElement {
+    }
+    var HTMLMvxWalletConnectAppStoreIconElement: {
+        prototype: HTMLMvxWalletConnectAppStoreIconElement;
+        new (): HTMLMvxWalletConnectAppStoreIconElement;
+    };
     interface HTMLMvxWalletConnectDownloadElement extends Components.MvxWalletConnectDownload, HTMLStencilElement {
     }
     var HTMLMvxWalletConnectDownloadElement: {
         prototype: HTMLMvxWalletConnectDownloadElement;
         new (): HTMLMvxWalletConnectDownloadElement;
+    };
+    interface HTMLMvxWalletConnectGooglePlayIconElement extends Components.MvxWalletConnectGooglePlayIcon, HTMLStencilElement {
+    }
+    var HTMLMvxWalletConnectGooglePlayIconElement: {
+        prototype: HTMLMvxWalletConnectGooglePlayIconElement;
+        new (): HTMLMvxWalletConnectGooglePlayIconElement;
     };
     interface HTMLMvxWalletConnectScanElementEventMap {
         "downloadClick": any;
@@ -1289,7 +1316,10 @@ declare global {
         "mvx-unlock-panel-group": HTMLMvxUnlockPanelGroupElement;
         "mvx-unlock-provider-button": HTMLMvxUnlockProviderButtonElement;
         "mvx-wallet-connect": HTMLMvxWalletConnectElement;
+        "mvx-wallet-connect-app-gallery-icon": HTMLMvxWalletConnectAppGalleryIconElement;
+        "mvx-wallet-connect-app-store-icon": HTMLMvxWalletConnectAppStoreIconElement;
         "mvx-wallet-connect-download": HTMLMvxWalletConnectDownloadElement;
+        "mvx-wallet-connect-google-play-icon": HTMLMvxWalletConnectGooglePlayIconElement;
         "mvx-wallet-connect-scan": HTMLMvxWalletConnectScanElement;
         "mvx-wallet-provider-icon": HTMLMvxWalletProviderIconElement;
         "mvx-xalias-provider-icon": HTMLMvxXaliasProviderIconElement;
@@ -1725,7 +1755,16 @@ declare namespace LocalJSX {
          */
         "qrCodeSvg"?: string;
     }
+    interface MvxWalletConnectAppGalleryIcon {
+        "class"?: string;
+    }
+    interface MvxWalletConnectAppStoreIcon {
+        "class"?: string;
+    }
     interface MvxWalletConnectDownload {
+        "class"?: string;
+    }
+    interface MvxWalletConnectGooglePlayIcon {
         "class"?: string;
     }
     interface MvxWalletConnectScan {
@@ -1832,7 +1871,10 @@ declare namespace LocalJSX {
         "mvx-unlock-panel-group": MvxUnlockPanelGroup;
         "mvx-unlock-provider-button": MvxUnlockProviderButton;
         "mvx-wallet-connect": MvxWalletConnect;
+        "mvx-wallet-connect-app-gallery-icon": MvxWalletConnectAppGalleryIcon;
+        "mvx-wallet-connect-app-store-icon": MvxWalletConnectAppStoreIcon;
         "mvx-wallet-connect-download": MvxWalletConnectDownload;
+        "mvx-wallet-connect-google-play-icon": MvxWalletConnectGooglePlayIcon;
         "mvx-wallet-connect-scan": MvxWalletConnectScan;
         "mvx-wallet-provider-icon": MvxWalletProviderIcon;
         "mvx-xalias-provider-icon": MvxXaliasProviderIcon;
@@ -1921,7 +1963,10 @@ declare module "@stencil/core" {
             "mvx-unlock-panel-group": LocalJSX.MvxUnlockPanelGroup & JSXBase.HTMLAttributes<HTMLMvxUnlockPanelGroupElement>;
             "mvx-unlock-provider-button": LocalJSX.MvxUnlockProviderButton & JSXBase.HTMLAttributes<HTMLMvxUnlockProviderButtonElement>;
             "mvx-wallet-connect": LocalJSX.MvxWalletConnect & JSXBase.HTMLAttributes<HTMLMvxWalletConnectElement>;
+            "mvx-wallet-connect-app-gallery-icon": LocalJSX.MvxWalletConnectAppGalleryIcon & JSXBase.HTMLAttributes<HTMLMvxWalletConnectAppGalleryIconElement>;
+            "mvx-wallet-connect-app-store-icon": LocalJSX.MvxWalletConnectAppStoreIcon & JSXBase.HTMLAttributes<HTMLMvxWalletConnectAppStoreIconElement>;
             "mvx-wallet-connect-download": LocalJSX.MvxWalletConnectDownload & JSXBase.HTMLAttributes<HTMLMvxWalletConnectDownloadElement>;
+            "mvx-wallet-connect-google-play-icon": LocalJSX.MvxWalletConnectGooglePlayIcon & JSXBase.HTMLAttributes<HTMLMvxWalletConnectGooglePlayIconElement>;
             "mvx-wallet-connect-scan": LocalJSX.MvxWalletConnectScan & JSXBase.HTMLAttributes<HTMLMvxWalletConnectScanElement>;
             "mvx-wallet-provider-icon": LocalJSX.MvxWalletProviderIcon & JSXBase.HTMLAttributes<HTMLMvxWalletProviderIconElement>;
             "mvx-xalias-provider-icon": LocalJSX.MvxXaliasProviderIcon & JSXBase.HTMLAttributes<HTMLMvxXaliasProviderIconElement>;